Remove deprecated social_profile access hook
The access hook provided access to profiles in case a profile was set as
featured content on a landing page or dashboard. This ensured that those
pages look correct for anonymous users that may be able to view them.

The changed implementation allows anonymous users to view profiles
(since this is needed for e.g. the profile image or name) so this access
override is no longer needed. Instead we now protect data on a field
level.