Add SocialProfileSettingsForm test coverage
This adds a test to check that the permissions for the
SocialProfileSettingsForm are set correctly and that all the different
fields correspond to the correct permissions and settings.

While writing this test we're also forced to add more accessible labels
to the individual fields to make sure we can easily select the right
one.